Adjust previous 2017 date to 2018
Small change: adjusted date and var name on line 103 ($APR17 --> $APR18) and var references/comments on lines 121, 122, 149, 151 and 152.

This is to accommodate the changes to HEFCE OA policy released Nov 2016. See below for details:

"HEFCE published an update to its OA policy (http://www.hefce.ac.uk/pubs/year/2016/201635/)
As previously known, 01/04/2016 was not envisaged any longer the hard deadline ( to make submitted output OA  within three months of acceptance) , but now there is a para ( page 5, 19  ‘on deposit requirements’) stating 
 
To take account of the need for systems to be developed to support deposit-on-acceptance, during the first two years of the policy (1 April 2016 – 1 April 2018), outputs can be deposited up to three months after the date of publication. This flexibility will be subject to a review of the readiness of systems within the sector in autumn 2017. 
 
Date of publication is defined as first availability online (not the print date)
 
This basically is a relaxation on ‘acceptance’ (but not a let-off) for publications published 01/04/2016 – 01/04/2018. OA still needs to be achieved three months AFTER publication/availablity online."